Extract index page to .html file and update styles
